Refactor: deprecate v1 staff controller
